Refactor circular dependency chain
